Backend/requirements.txt
absl-py==0.11.0
appdirs==1.4.4
astunparse==1.6.3
black==20.8b1
blis==0.7.4
cachetools==4.2.1
catalogue==2.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
contextvars==2.4
cymem==2.0.5
dataclasses==0.8
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.0.0/es_core_news_sm-3.0.0-py3-none-any.whl
Flask==1.1.2
Flask-Cors==3.0.10
flatbuffers==1.12
gast==0.3.3
google-auth==1.26.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
immutables==0.15
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
murmurhash==1.0.5
mypy-extensions==0.4.3
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pathspec==0.8.1
pathy==0.3.6
preshed==3.0.5
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.3
pyparsing==2.4.7
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
smart-open==3.0.0
spacy==3.0.1
spacy-legacy==3.0.1
srsly==2.4.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
thinc==8.0.1
toml==0.10.2
tqdm==4.56.2
typed-ast==1.4.2
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wasabi==0.8.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
